+++ src/nrepl/project-types.ts
@@ -818,6 +818,30 @@ const projectTypes: { [id: string]: ProjectType } = {
       };
     },
   },
+  epupp: {
+    name: 'epupp',
+    cljsTypes: [],
+    cmd: [],
+    winCmd: [],
+    processShellUnix: true,
+    processShellWin: true,
+    useWhenExists: [],
+    defaultFallbackPort: 3339,
+    defaultNReplPortFile: ['.epupp-nrepl-port'],
+    defaultReplSessionNames: { primary: 'epupp' },
+    defaultFilePatterns: {
+      primary: {
+        'always-claim': [
+          '**/userscripts/**/*.cljs',
+          '**/epupp/**/*.cljs',
+          '**/live-tampers/**/*.cljs',
+          '**/tampers/**/*.cljs',
+        ],
+        'is-fallback-for': ['**/*.cljs'],
+      },
+    },
+    commandLine: undefined,
+  },
   'cljs-only': {
     name: 'cljs-only',
     cljsTypes: [],
@@ -1071,8 +1095,6 @@ export function getProjectTypeForName(name: string) {
 export async function detectProjectTypes(): Promise<string[]> {
   const rootUri = state.getProjectRootUri();
   const cljProjTypes = [
-    'custom',
-    'generic',
     'clj-projectless',
     'cljs-only',
     'babashka',
@@ -1080,20 +1102,25 @@ export async function detectProjectTypes(): Promise<string[]> {
     'nbb',
     'joyride',
     'scittle',
+    'squint',
+    'epupp',
+    'custom',
+    'generic',
   ];
+  const projectCandidates = [];
   for (const clj in projectTypes) {
     for (const projectFileName of projectTypes[clj].useWhenExists) {
       try {
         const uri = vscode.Uri.joinPath(rootUri, projectFileName);
         await vscode.workspace.fs.readFile(uri);
-        cljProjTypes.push(clj);
+        projectCandidates.push(clj);
         break;
       } catch {
         // this just means the file doesn't exist
       }
     }
   }
-  return cljProjTypes;
+  return [...projectCandidates, ...cljProjTypes];
 }
 
 export function getAllProjectTypes(): string[] {