+++ test-data/projects/epupp-mini/live-tampers/pez/foo.cljs
@@ -1,0 +1,1 @@
+(js/console.log "Hello from Epupp")